requirements.txt
pandas>=2.0
numpy>=1.25
matplotlib>=3.7
pydantic>=2.5
PyYAML>=6.0
python-dotenv>=1.0.0
ccxt>=4.3.79
ta>=0.11.0
schedule>=1.2.1
flask>=2.3.0
MetaTrader5
pyyaml